Moved java files in request package
It exists many dependencies between the main java and all the requests classes/methods, I had to change some classes and method to public from package-private due to this. We'll see if we can come back to something less permissive
